Refactoring of the code to move stuff to common.c in preparation for creating fastasequant
