+++ CMakeLists.txt
@@ -1,14 +1,23 @@
 cmake_minimum_required(VERSION 3.10)
 project(fastdemux)
 
+
+## For the debug build:
+## cmake -S . -B build -DCMAKE_BUILD_TYPE=Debug
+
+# Set default build type if not specified
+if(NOT CMAKE_BUILD_TYPE)
+  set(CMAKE_BUILD_TYPE Release)
+endif()
+
 # Set C standard
 set(CMAKE_C_STANDARD 99)
 
-# Find zlib
+# Find zlib and OpenMP
 find_package(ZLIB REQUIRED)
 find_package(OpenMP REQUIRED)
 
-# Include directories for htslib and klib
+# Include directories for klib macro library
 include_directories(${CMAKE_SOURCE_DIR}/klib)
 
 # Set paths to htslib library and include directories
@@ -19,11 +28,15 @@ link_directories(/wsu/el7/groups/piquelab/samtools/1.19/lib)
 # Include zlib headers
 include_directories(${ZLIB_INCLUDE_DIRS})
 
-# Your source files
-add_executable(${PROJECT_NAME} main.c)
+# Define executables
+add_executable(fastdemux fastdemux.c common.c)
+##add_executable(fastqase fastqase.c common.c)
+
 
 # Link htslib
-target_link_libraries(${PROJECT_NAME} PUBLIC hts m OpenMP::OpenMP_C ${ZLIB_LIBRARIES})
+target_link_libraries(fastdemux PUBLIC hts m OpenMP::OpenMP_C ${ZLIB_LIBRARIES})
+##target_link_libraries(fastqase PUBLIC hts m OpenMP::OpenMP_C ${ZLIB_LIBRARIES})
+
 
 # Set Debug and Release configuration flags
 set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -pg -g")
@@ -32,4 +45,6 @@ set(CMAKE_C_FLAGS_RELEASE "${CMAKE_C_FLAGS_RELEASE} -O3")
 set(CMAKE_INSTALL_PREFIX "/wsu/home/groups/piquelab/apps/el7/misc/")
 
 # Specify the install rules
-install(TARGETS ${PROJECT_NAME} DESTINATION bin)
+install(TARGETS fastdemux DESTINATION bin)
+# Specify the install rules
+##install(TARGETS fastdemux fastqase DESTINATION bin)
